refactoring: incapsulated event subscription in TonConnect class & moved types and util functions
